requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
tqdm
Pillow
scikit-learn